+++ src/idunno.Bluesky/Graph/BlueskyServer.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) Barry Dorrans. All rights reserved.
+// Copyright (c) Barry Dorrans. All rights reserved.
 // Licensed under the MIT License.
 
 using System.Diagnostics.CodeAnalysis;
@@ -40,6 +40,9 @@ public static partial class BlueskyServer
 
     // https://docs.bsky.app/docs/api/app-bsky-graph-get-lists
     private const string GetListsEndpoint = "/xrpc/app.bsky.graph.getLists";
+
+    // https://docs.bsky.app/docs/api/app-bsky-graph-get-lists-with-membership
+    private const string GetListsWithMembershipEndpoint = "/xrpc/app.bsky.graph.getListsWithMembership";
 
     // https://docs.bsky.app/docs/api/app-bsky-graph-get-mutes
     private const string GetMutesEndpoint = "/xrpc/app.bsky.graph.getMutes";
@@ -327,7 +330,7 @@ public static partial class BlueskyServer
             $"{GetFollowsEndpoint}?{queryString}",
             credentials: accessCredentials,
             httpClient: httpClient,
-            jsonSerializerOptions : BlueskyJsonSerializerOptions,
+            jsonSerializerOptions: BlueskyJsonSerializerOptions,
             onCredentialsUpdated: onCredentialsUpdated,
             subscribedLabelers: subscribedLabelers,
             cancellationToken: cancellationToken).ConfigureAwait(false);
@@ -808,6 +811,106 @@ public static partial class BlueskyServer
     }
 
     /// <summary>
+    /// Enumerates the lists created by the session user, and includes membership information about actor in those lists.
+    /// Only supports curation and moderation lists (no reference lists, used in starter packs). Requires auth.
+    /// </summary>
+    /// <param name="actor">The account (actor) to check for membership.</param>
+    /// <param name="limit">The maximum number of lists that should be return in a page.</param>
+    /// <param name="cursor">An optional cursor for pagination.</param>
+    /// <param name="purposes">Optional filter by list purpose. If not specified, all supported types are returned. Possible values: [modlist, curatelist]</param>
+    /// <param name="service">The <see cref="Uri"/> of the service to retrieve the lists from.</param>
+    /// <param name="accessCredentials">The <see cref="AccessCredentials"/> used to authenticate to <paramref name="service"/>.</param>
+    /// <param name="httpClient">An <see cref="HttpClient"/> to use when making a request to the <paramref name="service"/>.</param>
+    /// <param name="onCredentialsUpdated">An <see cref="Action{T}" /> to call if the credentials in the request need updating.</param>
+    /// <param name="loggerFactory">An instance of <see cref="ILoggerFactory"/> to use to create a logger.</param>
+    /// <param name="subscribedLabelers">An optional list of <see cref="Did"/>s of labelers to retrieve labels applied to the post view.</param>
+    /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+    /// <returns>The task object representing the asynchronous operation.</returns>
+    /// <exception cref="ArgumentException">Thrown when <paramref name="actor"/> is <see langword="null"/> or empty or whitespace.</exception>
+    /// <exception cref="ArgumentNullException">Thrown when <paramref name="service"/>, <paramref name="accessCredentials"/>, or <paramref name="httpClient"/> is <see langword="null"/>.</exception>
+    /// <exception cref="ArgumentOutOfRangeException">Thrown when <paramref name="limit"/> is &lt; 1 or &gt; 100.</exception>
+    [UnconditionalSuppressMessage(
+        "Trimming",
+        "IL2026:Members annotated with 'RequiresUnreferencedCodeAttribute' require dynamic access otherwise can break functionality when trimming application code",
+        Justification = "All types are preserved in the JsonSerializerOptions call to Get().")]
+    [UnconditionalSuppressMessage("AOT",
+        "IL3050:Calling members annotated with 'RequiresDynamicCodeAttribute' may break functionality when AOT compiling.",
+        Justification = "All types are preserved in the JsonSerializerOptions call to Get().")]
+    public static async Task<AtProtoHttpResult<PagedViewReadOnlyCollection<ListWithMembership>>> GetListsWithMembership(
+        AtIdentifier actor,
+        int? limit,
+        string? cursor,
+        string[]? purposes,
+        Uri service,
+        AccessCredentials accessCredentials,
+        HttpClient httpClient,
+        Action<AtProtoCredential>? onCredentialsUpdated = null,
+        ILoggerFactory? loggerFactory = default,
+        IEnumerable<Did>? subscribedLabelers = null,
+        CancellationToken cancellationToken = default)
+    {
+        ArgumentException.ThrowIfNullOrWhiteSpace(actor);
+
+        ArgumentNullException.ThrowIfNull(service);
+        ArgumentNullException.ThrowIfNull(accessCredentials);
+        ArgumentNullException.ThrowIfNull(httpClient);
+        if (limit is not null)
+        {
+            ArgumentOutOfRangeException.ThrowIfLessThan((int)limit, 1);
+            ArgumentOutOfRangeException.ThrowIfGreaterThan((int)limit, 100);
+        }
+
+        StringBuilder queryStringBuilder = new();
+        queryStringBuilder.Append(CultureInfo.InvariantCulture, $"actor={Uri.EscapeDataString(actor!)}");
+        if (limit is not null)
+        {
+            queryStringBuilder.Append(CultureInfo.InvariantCulture, $"&limit={limit}");
+        }
+        if (cursor is not null)
+        {
+            queryStringBuilder.Append(CultureInfo.InvariantCulture, $"&cursor={Uri.EscapeDataString(cursor)}");
+        }
+        if (purposes is not null)
+        {
+            foreach (string purpose in purposes)
+            {
+                queryStringBuilder.Append(CultureInfo.InvariantCulture, $"&purposes={Uri.EscapeDataString(purpose)}");
+
+            }
+        }
+        string queryString = queryStringBuilder.ToString();
+
+        AtProtoHttpClient<GetListsWithMembershipResponse> client = new(AppViewProxy, loggerFactory);
+        AtProtoHttpResult<GetListsWithMembershipResponse> response = await client.Get(
+            service,
+            $"{GetListsWithMembershipEndpoint}?{queryString}",
+            credentials: accessCredentials,
+            httpClient: httpClient,
+            jsonSerializerOptions: BlueskyJsonSerializerOptions,
+            onCredentialsUpdated: onCredentialsUpdated,
+            subscribedLabelers: subscribedLabelers,
+            cancellationToken: cancellationToken).ConfigureAwait(false);
+        if (response.Succeeded)
+        {
+            return new AtProtoHttpResult<PagedViewReadOnlyCollection<ListWithMembership>>(
+                new PagedViewReadOnlyCollection<ListWithMembership>(response.Result.ListsWithMembership, response.Result.Cursor),
+                response.StatusCode,
+                response.HttpResponseHeaders,
+                response.AtErrorDetail,
+                response.RateLimit);
+        }
+        else
+        {
+            return new AtProtoHttpResult<PagedViewReadOnlyCollection<ListWithMembership>>(
+                null,
+                response.StatusCode,
+                response.HttpResponseHeaders,
+                response.AtErrorDetail,
+                response.RateLimit);
+        }
+    }
+
+    /// <summary>
     /// Get a paged list of muted profiles for the current user. Requires authentication.
     /// </summary>
     /// <param name="limit">The maximum number of lists that should be return in a page.</param>
@@ -826,7 +929,8 @@ public static partial class BlueskyServer
         "Trimming",
         "IL2026:Members annotated with 'RequiresUnreferencedCodeAttribute' require dynamic access otherwise can break functionality when trimming application code",
         Justification = "All types are preserved in the JsonSerializerOptions call to Get().")]
-    [UnconditionalSuppressMessage("AOT",
+    [UnconditionalSuppressMessage(
+        "AOT",
         "IL3050:Calling members annotated with 'RequiresDynamicCodeAttribute' may break functionality when AOT compiling.",
         Justification = "All types are preserved in the JsonSerializerOptions call to Get().")]
     public static async Task<AtProtoHttpResult<PagedViewReadOnlyCollection<ProfileView>>> GetMutes(